Rework the disassembly interface in advance of AArch64 disassembly.
We now have a number of changes to the internal interfaces for the
disassembly, to allow us to disassemble more than just ARM and Thumb.
This means that many of the places where we passed a flag, we now
pass a structure pointer which contains information about the
architecture.

ARM and Thumb seem to work, but AArch64 is currently stubbed.