Added support for disassembling AArch64 or ARM, or Thumb.
We can now disassemble AArch64, ARM and Thumb with the DisassembleArch
or *MemoryI or *DumpI. The Debugger_Disassemble SWI now uses the default
architecture so in RISC OS 64 it will disassemble 64bit code.

The disassembly for AArch64 doesn't seem quite right when disassembling
hello_world64 with *DumpI. It's not clear why - maybe this is a memory
lifetime issue. This can be looked at later.

The AArch64 isn't capitalised or formatted cleanly yet.