Initial implementation of the disassembly of AArch64 immediate constants.
The immediate constants are partially decoded from the instructions.
There are a few times when they're not right, because the instruction
doesn't make sense (SVC and UDF are obvious cases), and others where
it's printed as positive when it was meant to be negative. That's
not too hard to identify, as the types are given using an enumeration
that we are currently ignoring.